Fixed a problem where there where two similar question types (multiple choice and multi answer, which served the same purposes. I renamed MultipleChoice to MultipleChoiceWithMultipleAnswers, which better fits the assignment requirements and the purpose of a class. I renamed MultiAnswerQuestion to MultipleChoiceQuestion and made it so that only 1 answer is correct and user can also select only 1 answer. I adjusted all the classes accordingly, so there are no bugs.
