+++ src/main/java/models/QuestionType.java
@@ -9,8 +9,7 @@ public enum QuestionType {
     MULTIPLE_CHOICE("MULTIPLE_CHOICE"),
     PICTURE_RESPONSE("PICTURE_RESPONSE"),
     QUESTION_RESPONSE("QUESTION_RESPONSE"),
-    MULTI_ANSWER_ORDERED("MULTI_ANSWER_ORDERED"),
-    MULTI_ANSWER_UNORDERED("MULTI_ANSWER_UNORDERED");
+    MULTIPLE_CHOICE_WITH_MULTIPLE_ANSWERS("MULTIPLE_CHOICE_WITH_MULTIPLE_ANSWERS");
 
 
     private final String databaseValue;
